requirements.txt
# Core dependencies
pydicom>=2.4.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0

# Logging and utilities
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
